refactor: tidied up project directory structure, added alias for importing modules from specific folders
