+++ src/lib/transientInput.js
