+++ vite.config.js
@@ -1,11 +1,19 @@
 import { defineConfig } from 'vite';
+import path from 'node:path';
 
 export default defineConfig({
-  root: 'src/renderer',                  
+  root: 'src/',
+
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src/'),
+    },
+  },
+
   server: {
     port: 3000,
     strictPort: true,
-    open: false
+    open: false,
   },
   build: {
     outDir: '../dist',
